Move mixin to separate file
